Add files via upload
This folder is for the EEG/Emotion detection element of the project. "print_raw.py" generates a data stream from the EEG headset. The .csv files in "Training Data" are sample EEG recordings taken while attempting to invoke a specific emotional response on the user. The "tensor_test.py" is an attempt to create a trained tensorflow classifier based on the "training.csv" file.